+++ packages/toolbar/src/browser/toolbar.tsx
@@ -129,8 +129,8 @@ export class ToolbarImpl extends TabBarToolbar {
         return args;
     }
 
-    protected renderGroupsInColumn(groups: ToolbarItem[][], alignment: ToolbarAlignment): React.ReactNode {
-        const nodes: React.ReactNodeArray = [];
+    protected renderGroupsInColumn(groups: ToolbarItem[][], alignment: ToolbarAlignment): React.ReactNode[] {
+        const nodes: React.ReactNode[] = [];
         groups.forEach((group, groupIndex) => {
             if (nodes.length && group.length) {
                 nodes.push(<div key={`toolbar-separator-${groupIndex}`} className='separator' />);